requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.22
instagrapi==1.19.4
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.4.2